Changed definition of base flow class
tendrl-spec: refactor_common_code_and_move
tendrl-bug-id:Tendrl/common#62
tendrl-bug-id:Tendrl/common#60
tendrl-bug-id:Tendrl/common#82
tendrl-bug-id:Tendrl/common#83
tendrl-bug-id:Tendrl/common#84
