+++ tendrl/common/config.py
@@ -13,13 +13,12 @@ class TendrlConfig(ConfigParser.SafeConfigParser):
     def __init__(self, module, cfg_file_path):
         ConfigParser.SafeConfigParser.__init__(self)
 
-        self.module = module
         self.path = cfg_file_path
 
         if not os.path.exists(self.path):
             err = ConfigNotFound(
                 "Configuration for module: %s not found at %s" %
-                self.module, self.path
+                module, self.path
             )
             LOG.error(err, exc_info=True)
             raise err
